LayoutTests/http/tests/inspector/network/network-content-replacement-xhr-expected.txt

CONSOLE MESSAGE: line 39: Done.
Tests NetworkResourcesData logic for XHR content replacement.

http://127.0.0.1:8000/inspector/network/resources/resource.php?size=200
resource.content: 
http://127.0.0.1:8000/inspector/network/resources/resource.php?size=100
resource.content: ****************************************************************************************************
http://127.0.0.1:8000/inspector/network/resources/resource.php?size=201
resource.content: 
http://127.0.0.1:8000/inspector/network/resources/resource.php?size=100
resource.content: ****************************************************************************************************
